Custom network serialization for UpdateMapIncremental. Unit tests for custom network serialization.  Minor cleanup in engine/map/server

+++ dune-server/src/main/java/lv/k2611a/service/game/GameServiceImpl.java
@@ -236,8 +236,8 @@ public class GameServiceImpl implements GameService {
             int playerId = clientConnection.getPlayerId();
             Player player = map.getPlayerById(playerId);
             UpdateMoney updateMoney = new UpdateMoney();
-            updateMoney.setMoney(player.getMoney());
-            updateMoney.setElectricity(player.getElectricity());
+            updateMoney.setMoney((int) player.getMoney());
+            updateMoney.setElectricity((int) player.getElectricity());
             clientConnection.sendMessage(updateMoney);
         }
     }
